config

+++ config/pcs.php
@@ -1,5 +1,5 @@
 <?php
 
 return [
-    'team_id' => (int) env('PCS_TEAM_ID', 1),
+    'team_id' => env('PCS_TEAM_ID', 1),
 ];
